hotfix db connection, fixed controller and model
